Refactor web appData modules
