+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FilterSupport.kt
@@ -166,7 +166,7 @@ fun matchesCardFilter(filter: CardFilter, card: CardSummary): Boolean {
 }
 
 fun matchesDeckFilterDefinition(filterDefinition: DeckFilterDefinition, card: CardSummary): Boolean {
-    // Keep deck matching semantics aligned with apps/web/src/appData/domain.ts::matchesDeckFilterDefinition and apps/ios/Flashcards/Flashcards/CardFilterSupport.swift::matchesDeckFilterDefinition.
+    // Keep deck matching semantics aligned with apps/web/src/appData/domain/index.ts::matchesDeckFilterDefinition and apps/ios/Flashcards/Flashcards/Cards/CardFilterSupport.swift::matchesDeckFilterDefinition.
     if (filterDefinition.effortLevels.isNotEmpty() && filterDefinition.effortLevels.contains(card.effortLevel).not()) {
         return false
     }
