+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/ReviewSupport.kt
@@ -3,7 +3,7 @@ package com.flashcardsopensourceapp.data.local.model
 // Keep review queue ordering aligned with:
 // - apps/ios/Flashcards/Flashcards/Review/ReviewQuerySupport.swift::compareCardsForReviewOrder
 // - apps/ios/Flashcards/Flashcards/Database/CardStore/CardStore+ReadSQL.swift review queue ORDER BY
-// - apps/web/src/appData/domain.ts::compareCardsForReviewOrder
+// - apps/web/src/appData/domain/index.ts::compareCardsForReviewOrder
 // Active queue contract: recent due cards within the inclusive one-hour window first,
 // then older due cards, then null due/new cards. Future cards are excluded from the
 // active queue and can appear later in timeline ordering; malformed dueAt values are
