+++ apps/ios/Flashcards/Flashcards/Database/CardStore/CardStore+ReadSQL.swift
@@ -284,7 +284,7 @@ extension CardStore {
         // Keep review queue ordering aligned with:
         // - apps/ios/Flashcards/Flashcards/Review/ReviewQuerySupport.swift::compareCardsForReviewOrder
         // - apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/ReviewSupport.kt::sortCardsForReviewQueue
-        // - apps/web/src/appData/domain.ts::compareCardsForReviewOrder
+        // - apps/web/src/appData/domain/index.ts::compareCardsForReviewOrder
         // Ordering contract: recent due cards within the inclusive one-hour window first, then older due cards,
         // then nil dueAt new cards. Future and malformed dueAt values are excluded from the active queue.
         // If this changes, mirror the same change across all three clients in the same change.
