+++ apps/web/src/appData/domain/index.ts
@@ -1,6 +1,6 @@
 import type {
   ReviewableCardScheduleState,
-} from "../../../backend/src/schedule";
+} from "../../../../backend/src/schedule";
 import { canonicalizeDueAtForSync, parseDueAtMillis } from "./dueAt";
 import type {
   Card,
@@ -18,8 +18,8 @@ import type {
   WorkspaceTagSummary,
   WorkspaceTagsSummary,
   WorkspaceSummary,
-} from "../types";
-import { ALL_CARDS_DECK_LABEL } from "../deckFilters";
+} from "../../types";
+import { ALL_CARDS_DECK_LABEL } from "../../deckFilters";
 
 type LastWriteWinsRecord = Readonly<{
   clientUpdatedAt: string;
@@ -143,7 +143,7 @@ function findMatchingTag(tags: ReadonlyArray<string>, requestedTag: string): str
   return tags.find((tag) => normalizeTagKey(tag) === requestedTagKey)?.trim() ?? null;
 }
 
-/** Keep deck matching semantics aligned with apps/ios/Flashcards/Flashcards/CardFilterSupport.swift and apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FilterSupport.kt: effort is inclusive and tags match on any overlap. */
+/** Keep deck matching semantics aligned with apps/ios/Flashcards/Flashcards/Cards/CardFilterSupport.swift and apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FilterSupport.kt: effort is inclusive and tags match on any overlap. */
 export function matchesDeckFilterDefinition(filterDefinition: DeckFilterDefinition, card: Card): boolean {
   if (filterDefinition.effortLevels.length > 0 && filterDefinition.effortLevels.includes(card.effortLevel) === false) {
     return false;
