+++ apps/web/src/appData/session/warmStart.ts
@@ -1,5 +1,5 @@
-import { isBrowserReauthRequired } from "../accountDeletion";
-import type { SessionInfo, WorkspaceSummary } from "../types";
+import { isBrowserReauthRequired } from "../../accountDeletion";
+import type { SessionInfo, WorkspaceSummary } from "../../types";
 
 export type SessionVerificationState = "unverified" | "verified";
