+++ apps/web/src/screens/review/ReviewScreen.tsx
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import ReactMarkdown from "react-markdown";
 import { useAppData, useReviewProgressBadge } from "../../appData";
 import { ALL_CARDS_REVIEW_FILTER, currentReviewCard, isCardDue } from "../../appData/domain";
-import { parseDueAtMillis } from "../../appData/dueAt";
+import { parseDueAtMillis } from "../../appData/domain/dueAt";
 import { useI18n } from "../../i18n";
 import { captureAppOperationError } from "../../observability/appOperationObservation";
 import { normalizeCaughtError } from "../../observability/webObservability";
