+++ apps/web/src/screens/review/ReviewScreenTestSupport.tsx
@@ -5,7 +5,7 @@ import ReactDOM from "react-dom/client";
 import { MemoryRouter } from "react-router-dom";
 import { afterEach, beforeEach, vi } from "vitest";
 import { I18nProvider } from "../../i18n";
-import type { AppDataContextValue } from "../../appData/types";
+import type { AppDataContextValue } from "../../appData";
 import { clearLoadingSnapshotFallbackStorage } from "../shared/loadingSnapshots";
 import type {
   Card,
